backend/requirements.txt
fastapi==0.111.0
uvicorn==0.29.0
pydantic==2.7.4
pydantic-settings==2.2.1
SQLAlchemy==2.0.30
alembic==1.13.1
python-multipart==0.0.9
python-jose==3.3.0
passlib[bcrypt]==1.7.4
email-validator==2.1.1
python-dotenv==1.0.1
starlette==0.36.3
APScheduler==3.10.4
schedule==1.2.0
httpx==0.27.0
dnspython==2.6.1
